requirements.txt
asgiref==3.5.2
autopep8==1.7.0
dj-database-url==1.0.0
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pycodestyle==2.9.1
pytz==2022.4
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
